Remove redundant type conversion

+++ internal/db/nulltypes.go
@@ -46,7 +46,7 @@ func NewNullFloat64(f *float64) sql.NullFloat64 {
 	if f == nil {
 		return sql.NullFloat64{Valid: false}
 	}
-	return sql.NullFloat64{Float64: float64(*f), Valid: true}
+	return sql.NullFloat64{Float64: *f, Valid: true}
 }
 
 func NewNullUUID(id *uuid.UUID) uuid.NullUUID {
